Teach the app to accept 'remark' arguments but do nothing
Add Remark to model component (copy from Address, remove the regex and
rename accordingly).

Add a RemarkCommand that takes in an Index and Remark for remark, and
print those two parameters as the error message.

Add RemarkCommandTest that will test that executeUndoableCommand()
throws an Exception and test the equals method.

Add RemarkCommandParser that will know how to parse two arguments, one
index and one with prefix 'r/'

Add RemarkCommandParserTest that will test different boundary values
for RemarkCommandParser.

Modify AddressBookParser to accept a RemarkCommand and use the newly
implemented RemarkCommandParser.

Add new test to AddressBookParserTest, which will test that typing
"remark" returns an instance of RemarkCommand and to ensure that what
the user inputs generates the correct command.

Add test for Remark (to test the equals method).